+++ ycc_hull/api/members.py
@@ -1,6 +1,7 @@
 """
 Member API endpoints.
 """
+
 from collections.abc import Sequence
 from datetime import date
 
@@ -19,16 +20,10 @@ controller = MembersController()
 
 @api_members.get("/api/v1/members")
 async def members_get(
-    year: int | None = None, user: User = Depends(auth)
+    year: int, user: User = Depends(auth)
 ) -> Sequence[MemberPublicInfoDto]:
-    current_year = date.today().year
-
-    if year != current_year and not (user.admin or user.committee_member):
-        raise create_http_exception_403(
-            f"You do not have permission to list members for {year}"
-        )
-
-    return await controller.find_all_public_infos(year)
+    _check_can_access_year(year, user)
+    return await controller.find_all_public_infos(year=year)
 
 
 @api_members.get("/api/v1/membership-types")
@@ -42,3 +37,12 @@ async def users_get(user: User = Depends(auth)) -> Sequence[UserDto]:
         raise create_http_exception_403("You do not have permission to list users")
 
     return await controller.find_all_users()
+
+
+def _check_can_access_year(year: int | None, user: User) -> None:
+    current_year = date.today().year
+
+    if year != current_year and not (user.admin or user.committee_member):
+        raise create_http_exception_403(
+            f"You do not have permission to view members for {year}"
+        )