+++ ycc_hull/controllers/helpers_controller.py
@@ -1,6 +1,7 @@
 """
 Helpers controller.
 """
+
 from collections.abc import Sequence
 from datetime import datetime
 
@@ -23,9 +24,12 @@ from ycc_hull.db.entities import (
 from ycc_hull.models.helpers_dtos import (
     HelperTaskCategoryDto,
     HelperTaskDto,
+    HelperTaskMarkAsDoneRequestDto,
     HelperTaskMutationRequestDto,
+    HelperTaskValidationRequestDto,
 )
 from ycc_hull.models.user import User
+from ycc_hull.utils import get_now
 
 
 class HelpersController(BaseController):
@@ -40,29 +44,40 @@ class HelpersController(BaseController):
         )
 
     async def find_all_tasks(
-        self, published: bool | None = None
+        self, *, year: int | None = None, published: bool | None = None
     ) -> Sequence[HelperTaskDto]:
-        return await self._find_tasks(None, published)
+        return await self._find_tasks(year=year, task_id=None, published=published)
 
     async def find_task_by_id(
-        self, task_id: int, published: bool | None = None
+        self,
+        task_id: int,
+        *,
+        published: bool | None = None,
+        session: AsyncSession | None = None,
     ) -> HelperTaskDto | None:
-        return await self._find_task_by_id(task_id, published)
+        return await self._find_task_by_id(
+            task_id, published=published, session=session
+        )
 
     async def get_task_by_id(
-        self, task_id: int, published: bool | None = None
+        self,
+        task_id: int,
+        *,
+        published: bool | None = None,
+        session: AsyncSession | None = None,
     ) -> HelperTaskDto:
-        task = await self.find_task_by_id(task_id, published)
+        task = await self.find_task_by_id(task_id, published=published, session=session)
         if task:
             return task
         raise ControllerNotFoundException("Task not found")
 
     async def create_task(
-        self, task_mutation_request: HelperTaskMutationRequestDto, user: User
+        self, request: HelperTaskMutationRequestDto, user: User
     ) -> HelperTaskDto:
+        # Admins/editors have full power (e.g., administer things happened in the past)
         async with self.database_context.async_session() as session:
             try:
-                task_entity = HelperTaskEntity(**task_mutation_request.model_dump())
+                task_entity = HelperTaskEntity(**request.model_dump())
                 session.add(task_entity)
                 await session.commit()
 
@@ -77,47 +92,52 @@ class HelpersController(BaseController):
                 return task
             except DatabaseError as exc:
                 raise self._handle_database_error(  # pylint: disable=raising-bad-type
-                    exc, "create task", user, task_mutation_request
+                    exc, what="create task", user=user, data=request
                 )
 
     async def update_task(
         self,
         task_id: int,
-        task_mutation_request: HelperTaskMutationRequestDto,
+        request: HelperTaskMutationRequestDto,
         user: User,
     ) -> HelperTaskDto:
+        # Admins/editors have full power (e.g., administer things happened in the past)
         async with self.database_context.async_session() as session:
             old_task = await self._get_task_by_id(task_id, session=session)
 
+            anyone_signed_up = old_task.captain or old_task.helpers
+            same_timing = (
+                request.starts_at == old_task.starts_at
+                and request.ends_at == old_task.ends_at
+                and request.deadline == old_task.deadline
+            )
+
             # Check: cannot change timing if anyone has signed up
-            if old_task.captain or old_task.helpers:
-                if (
-                    task_mutation_request.starts_at != old_task.starts_at
-                    or task_mutation_request.ends_at != old_task.ends_at
-                    or task_mutation_request.deadline != old_task.deadline
-                ):
-                    raise ControllerConflictException(
-                        "Cannot change timing after anyone has signed up"
-                    )
-                if not task_mutation_request.published:
-                    raise ControllerConflictException(
-                        "You must publish a task after anyone has signed up"
-                    )
+            if anyone_signed_up and not same_timing:
+                raise ControllerConflictException(
+                    "Cannot change timing after anyone has signed up"
+                )
+            # Check: must publish a task after anyone has signed up
+            if anyone_signed_up and not request.published:
+                raise ControllerConflictException(
+                    "You must publish a task after anyone has signed up"
+                )
+            # Check: cannot change timing after the task has been marked as done
+            if not same_timing and old_task.marked_as_done_at:
+                raise ControllerConflictException(
+                    "Cannot change timing after the task has been marked as done"
+                )
 
             # Check: if a captain has signed up then the new licence must be active for the captain
-            if (
-                old_task.captain
-                and task_mutation_request.captain_required_licence_info_id
-                != (
-                    old_task.captain_required_licence_info.id
-                    if old_task.captain_required_licence_info
-                    else None
-                )
+            if old_task.captain and request.captain_required_licence_info_id != (
+                old_task.captain_required_licence_info.id
+                if old_task.captain_required_licence_info
+                else None
             ):
                 captain_entity = old_task.captain.member.get_entity()
                 if not any(
                     licence_info_entity.infoid
-                    == task_mutation_request.captain_required_licence_info_id
+                    == request.captain_required_licence_info_id
                     for licence_info_entity in captain_entity.active_licence_infos
                 ):
                     raise ControllerConflictException(
@@ -125,7 +145,7 @@ class HelpersController(BaseController):
                     )
 
             # Check: cannot decrease helpers maximum count below signed up helpers count
-            if task_mutation_request.helper_max_count < len(old_task.helpers):
+            if request.helper_max_count < len(old_task.helpers):
                 raise ControllerConflictException(
                     "Cannot decrease helpers maximum count below signed up helpers count"
                 )
@@ -133,7 +153,7 @@ class HelpersController(BaseController):
             try:
                 task_entity = old_task.get_entity()
                 old_task = await HelperTaskDto.create(task_entity)
-                self._update_entity_from_dto(task_entity, task_mutation_request)
+                self._update_entity_from_dto(task_entity, request)
                 await session.commit()
 
                 new_task = await HelperTaskDto.create(task_entity)
@@ -151,14 +171,14 @@ class HelpersController(BaseController):
                 return new_task
             except DatabaseError as exc:
                 raise self._handle_database_error(  # pylint: disable=raising-bad-type
-                    exc, "update task", user, task_mutation_request
+                    exc, what="update task", user=user, data=request
                 )
 
     async def sign_up_as_captain(self, task_id: int, user: User) -> None:
         async with self.database_context.async_session() as session:
             task = await self._get_task_by_id(task_id, published=True, session=session)
 
-            await self._check_can_sign_up(task, user.member_id)
+            await self._check_can_sign_up(task=task, member_id=user.member_id)
             if task.captain:
                 raise ControllerConflictException("Task already has a captain")
 
@@ -169,15 +189,9 @@ class HelpersController(BaseController):
                         f"Task captain needs licence: {required_licence}"
                     )
 
-            task_entity = (
-                await session.scalars(
-                    select(HelperTaskEntity)
-                    .options(lazyload("*"))
-                    .where(HelperTaskEntity.id == task_id)
-                )
-            ).one()
+            task_entity = await self._get_task_entity_by_id(task_id, session=session)
             task_entity.captain_id = user.member_id
-            task_entity.captain_signed_up_at = datetime.now()
+            task_entity.captain_signed_up_at = get_now()
             await session.commit()
 
             session.add(
@@ -186,15 +200,15 @@ class HelpersController(BaseController):
             await session.commit()
 
     async def sign_up_as_helper(self, task_id: int, user: User) -> None:
-        task = await self.get_task_by_id(task_id, published=True)
-
-        await self._check_can_sign_up(task, user.member_id)
-        if len(task.helpers) >= task.helper_max_count:
-            raise ControllerConflictException("Task helper limit reached")
-
         async with self.database_context.async_session() as session:
+            task = await self.get_task_by_id(task_id, published=True, session=session)
+
+            await self._check_can_sign_up(task=task, member_id=user.member_id)
+            if len(task.helpers) >= task.helper_max_count:
+                raise ControllerConflictException("Task helper limit reached")
+
             helper = HelperTaskHelperEntity(
-                task_id=task.id, member_id=user.member_id, signed_up_at=datetime.now()
+                task_id=task.id, member_id=user.member_id, signed_up_at=get_now()
             )
             session.add(helper)
             await session.commit()
@@ -204,21 +218,106 @@ class HelpersController(BaseController):
             )
             await session.commit()
 
+    async def mark_as_done(
+        self, task_id: int, request: HelperTaskMarkAsDoneRequestDto, user: User
+    ) -> None:
+        async with self.database_context.async_session() as session:
+            task = await self._get_task_by_id(task_id, published=True, session=session)
+
+            if task.marked_as_done_at:
+                raise ControllerConflictException("Task already marked as done")
+
+            task_entity = await self._get_task_entity_by_id(task_id, session=session)
+            task_entity.marked_as_done_at = get_now()
+            task_entity.marked_as_done_by_id = user.member_id
+            task_entity.marked_as_done_comment = request.comment
+            await session.commit()
+
+            session.add(
+                create_audit_entry(
+                    user, f"Helpers/Tasks/MarkAsDone/{task_id}", {"request": request}
+                )
+            )
+            await session.commit()
+
+    async def validate(
+        self, task_id: int, request: HelperTaskValidationRequestDto, user: User
+    ) -> None:
+        async with self.database_context.async_session() as session:
+            task = await self._get_task_by_id(task_id, published=True, session=session)
+
+            if task.validated_at:
+                raise ControllerConflictException("Task already validated")
+
+            helper_ids_in_db = set(helper.member.id for helper in task.helpers)
+            helper_ids_in_request = set(
+                helper.member.id
+                for helper in request.helpers_to_validate + request.helpers_to_remove
+            )
+
+            if helper_ids_in_db != helper_ids_in_request:
+                raise ControllerConflictException(
+                    "Validation request is missing helpers"
+                )
+
+            to_be_removed = set(
+                helper.member.id for helper in request.helpers_to_remove
+            )
+
+            task_entity = await self._get_task_entity_by_id(task_id, session=session)
+            now = get_now()
+
+            if not task_entity.marked_as_done_at:
+                task_entity.marked_as_done_at = now
+                task_entity.marked_as_done_by_id = user.member_id
+
+            task_entity.validated_at = now
+            task_entity.validated_by_id = user.member_id
+            task_entity.validation_comment = request.comment
+
+            for helper_entity in await task_entity.awaitable_attrs.helpers:
+                if helper_entity.member_id in to_be_removed:
+                    await session.delete(helper_entity)
+
+            await session.commit()
+
+            session.add(
+                create_audit_entry(
+                    user,
+                    f"Helpers/Tasks/Validate/{task_id}",
+                    {"request": request},
+                )
+            )
+            await session.commit()
+
     async def _find_tasks(
         self,
+        *,
+        year: int | None,
         task_id: int | None,
         published: bool | None,
         session: AsyncSession | None = None,
     ) -> Sequence[HelperTaskDto]:
         query = select(HelperTaskEntity)
 
-        exclude_long_description: bool = task_id is None
+        exclude_large_fields: bool = task_id is None
 
-        if exclude_long_description:
+        if exclude_large_fields:
             query = query.options(
-                defer(HelperTaskEntity.long_description, raiseload=True)
+                defer(HelperTaskEntity.long_description, raiseload=True),
+                defer(HelperTaskEntity.marked_as_done_comment, raiseload=True),
+                defer(HelperTaskEntity.validation_comment, raiseload=True),
             )
 
+        if year is not None:
+            query = query.where(
+                func.coalesce(  # pylint: disable=not-callable
+                    HelperTaskEntity.starts_at, HelperTaskEntity.deadline
+                ).between(
+                    datetime(year, 1, 1, 0, 0, 0, 0),
+                    datetime(year, 12, 31, 23, 59, 59, 0),
+                )
+            )
         if task_id is not None:
             query = query.where(HelperTaskEntity.id == task_id)
         if published is not None:
@@ -233,9 +332,11 @@ class HelpersController(BaseController):
 
         return await self.database_context.query_all(
             query,
-            async_transformer=HelperTaskDto.create_without_long_description
-            if exclude_long_description
-            else HelperTaskDto.create,
+            async_transformer=(
+                HelperTaskDto.create_without_large_fields
+                if exclude_large_fields
+                else HelperTaskDto.create
+            ),
             unique=True,
             session=session,
         )
@@ -243,28 +344,51 @@ class HelpersController(BaseController):
     async def _find_task_by_id(
         self,
         task_id: int,
+        *,
         published: bool | None,
         session: AsyncSession | None = None,
     ) -> HelperTaskDto | None:
-        tasks = await self._find_tasks(task_id, published, session=session)
+        tasks = await self._find_tasks(
+            year=None, task_id=task_id, published=published, session=session
+        )
         return tasks[0] if tasks else None
 
     async def _get_task_by_id(
         self,
         task_id: int,
+        *,
         published: bool | None = None,
         session: AsyncSession | None = None,
     ) -> HelperTaskDto:
-        task = await self._find_task_by_id(task_id, published, session=session)
+        task = await self._find_task_by_id(
+            task_id, published=published, session=session
+        )
         if task:
             return task
-        raise ControllerNotFoundException("Task not found")
+        raise ControllerNotFoundException(
+            "Task not found or not published" if published else "Task not found"
+        )
 
-    async def _check_can_sign_up(self, task: HelperTaskDto, member_id: int) -> None:
+    async def _get_task_entity_by_id(
+        self, task_id: int, *, session: AsyncSession
+    ) -> HelperTaskEntity:
+        return (
+            await session.scalars(
+                select(HelperTaskEntity)
+                .options(lazyload("*"))
+                .where(HelperTaskEntity.id == task_id)
+            )
+        ).one()
+
+    async def _check_can_sign_up(self, *, task: HelperTaskDto, member_id: int) -> None:
         if not task.published:
             raise ControllerConflictException("Cannot sign up to an unpublished task")
+        if task.marked_as_done_at:
+            raise ControllerConflictException("Cannot sign up to a task marked as done")
+        if task.validated_at:
+            raise ControllerConflictException("Cannot sign up to a validated task")
 
-        now = datetime.now()
+        now = get_now()
         if (task.starts_at and task.starts_at < now) or (
             task.deadline and task.deadline < now
         ):