+++ ycc_hull/controllers/holidays_controller.py
@@ -1,6 +1,7 @@
 """
 Holidays controller.
 """
+
 from collections.abc import Sequence
 
 from sqlalchemy import select