+++ ycc_hull/models/helpers_dtos.py
@@ -1,6 +1,7 @@
 """
 Helpers API DTO classes.
 """
+
 from collections.abc import Sequence
 from datetime import datetime
 
@@ -63,24 +64,57 @@ class HelperTaskDto(CamelisedBaseModelWithEntity[HelperTaskEntity]):
     captain: "HelperTaskHelperDto | None"
     helpers: Sequence["HelperTaskHelperDto"]
 
-    @classmethod
-    async def create(cls, task: HelperTaskEntity) -> "HelperTaskDto":
-        return await cls._create(task, await task.awaitable_attrs.long_description)
+    marked_as_done_at: datetime | None
+    marked_as_done_by: MemberPublicInfoDto | None
+    marked_as_done_comment: str | None
+    validated_at: datetime | None
+    validated_by: MemberPublicInfoDto | None
+    validation_comment: str | None
+
+    @property
+    def year(self) -> int:
+        if self.starts_at:
+            return self.starts_at.year
+        if self.ends_at:
+            return self.ends_at.year
+        if self.deadline:
+            return self.deadline.year
+        raise ValueError("Missing timing")
 
     @classmethod
-    async def create_without_long_description(
+    async def create(cls, task: HelperTaskEntity) -> "HelperTaskDto":
+        return await cls._create(
+            task,
+            long_description=await task.awaitable_attrs.long_description,
+            marked_as_done_comment=await task.awaitable_attrs.marked_as_done_comment,
+            validation_comment=await task.awaitable_attrs.validation_comment,
+        )
+
+    @classmethod
+    async def create_without_large_fields(
         cls, task: HelperTaskEntity
     ) -> "HelperTaskDto":
-        return await cls._create(task, None)
+        return await cls._create(
+            task,
+            long_description=None,
+            marked_as_done_comment=None,
+            validation_comment=None,
+        )
 
     @staticmethod
     async def _create(
-        task: HelperTaskEntity, long_description: str | None
+        task: HelperTaskEntity,
+        *,
+        long_description: str | None,
+        marked_as_done_comment: str | None,
+        validation_comment: str | None,
     ) -> "HelperTaskDto":
         captain = await task.awaitable_attrs.captain
         captain_required_licence_info = (
             await task.awaitable_attrs.captain_required_licence_info
         )
+        marked_as_done_by = await task.awaitable_attrs.marked_as_done_by
+        validated_by = await task.awaitable_attrs.validated_by
 
         return HelperTaskDto(
             entity=task,
@@ -98,11 +132,11 @@ class HelperTaskDto(CamelisedBaseModelWithEntity[HelperTaskEntity]):
             ends_at=task.ends_at,
             deadline=task.deadline,
             urgent=task.urgent,
-            captain_required_licence_info=await LicenceInfoDto.create(
-                captain_required_licence_info
-            )
-            if captain_required_licence_info
-            else None,
+            captain_required_licence_info=(
+                await LicenceInfoDto.create(captain_required_licence_info)
+                if captain_required_licence_info
+                else None
+            ),
             helper_min_count=task.helper_min_count,
             helper_max_count=task.helper_max_count,
             published=task.published,
@@ -119,6 +153,18 @@ class HelperTaskDto(CamelisedBaseModelWithEntity[HelperTaskEntity]):
                 await HelperTaskHelperDto.create(helper)
                 for helper in await task.awaitable_attrs.helpers
             ],
+            marked_as_done_at=task.marked_as_done_at,
+            marked_as_done_by=(
+                await MemberPublicInfoDto.create(marked_as_done_by)
+                if marked_as_done_by
+                else None
+            ),
+            marked_as_done_comment=marked_as_done_comment,
+            validated_at=task.validated_at,
+            validated_by=(
+                await MemberPublicInfoDto.create(validated_by) if validated_by else None
+            ),
+            validation_comment=validation_comment,
         )
 
 
@@ -150,23 +196,51 @@ class HelperTaskMutationRequestDto(CamelisedBaseModel):
 
     @model_validator(mode="after")
     def check_timing(self) -> "HelperTaskMutationRequestDto":
-        valid_shift = (
-            self.starts_at
-            and self.ends_at
-            and not self.deadline
-            and self.starts_at < self.ends_at
-        )
-        valid_deadline = not self.starts_at and not self.ends_at and self.deadline
+        if self.starts_at and self.ends_at and not self.deadline:
+            # Shifts have extra conditions
+            if self.starts_at >= self.ends_at:
+                raise ValueError("Invalid timing: start time must be before end time")
+            if self.starts_at.year != self.ends_at.year:
+                raise ValueError(
+                    "Invalid timing: start and end time must be in the same year"
+                )
+        elif not self.starts_at and not self.ends_at and self.deadline:
+            # Nothing more to validate on one value
+            pass
+        else:
+            raise ValueError(
+                "Invalid timing: either specify both start and end time for a shift or a deadline for a task"
+            )
 
-        if valid_shift or valid_deadline:
-            return self
-        raise ValueError("Invalid timing")
+        return self
 
     @model_validator(mode="after")
     def check_helper_min_max_count(self) -> "HelperTaskMutationRequestDto":
         if 0 <= self.helper_min_count <= self.helper_max_count:
             return self
         raise ValueError("Invalid minimum/maximum helper count")
+
+
+class HelperTaskMarkAsDoneRequestDto(CamelisedBaseModel):
+    """
+    Mark as done request DTO for helper task.
+    """
+
+    comment: str | None
+
+
+class HelperTaskValidationRequestDto(CamelisedBaseModel):
+    """
+    Validation request DTO for helper task.
+    """
+
+    helpers_to_validate: list["HelperTaskHelperDto"] = Field(
+        description="List of helpers to validate"
+    )
+    helpers_to_remove: list["HelperTaskHelperDto"] = Field(
+        description="List of helpers to remove (e.g., no show)"
+    )
+    comment: str | None
 
 
 class HelperTaskHelperDto(CamelisedBaseModelWithEntity[HelperTaskHelperEntity]):